requirements.txt
mysql-connector==2.1.6
requests==2.14.2